refactor: split codegen type inference
